Fix E2E test and remove product improve page
